pyproject.toml
[project]
name = "noai-watermark"
version = "0.1.16"
description = "AI Metadata Toolkit for cloning, checking, cleaning, and watermark removal"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0.0",
    "piexif>=1.1.3",
    "torch>=2.0.0",
    "diffusers>=0.25.0",
    "transformers>=4.35.0",
    "accelerate>=0.25.0",
    "controlnet-aux>=0.0.9",
    "color-matcher",
    "safetensors",
]

[project.urls]
Homepage = "https://github.com/mertizci/noai-watermark"
Repository = "https://github.com/mertizci/noai-watermark"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[project.scripts]
noai-watermark = "noai_cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = ["src"]
sources = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
